Fix the search for processes with the proper vhost config when ServerName isn't set in every vhost or a module updates r->server->server_hostname dynamically (e.g., mod_vhost_cdb).
In r753578, the check for vhost was added to searches for available
processes so that only processes with the same configuration would 
be found.

(In what seems like a separate issue, it was also added to the
search for the count of matching processes (spawn control).)

The check was implemented by saving r->server->server_hostname
and comparing that value.  This fails in the following cases:

a) ServerName hasn't been configured in each vhost
b) some module updates r->server->server_hostname dynamically
   (mod_vhost_cdb is one such module)

Change the implementation to just use the server_rec.  This should
be sufficient for distinguishing potentially different vhost
configs, and it provides addressibility to server->server_hostname
for logging as well.


git-svn-id: http://svn.apache.org/repos/asf/httpd/mod_fcgid/trunk@939472 13f79535-47bb-0310-9956-ffa450edef68